feat: moved resource_importer to model_config and renamed model_config to resources.py
